canvas/requirements.txt
canvasapi
mistune
Pillow
pyyaml
requests
